+++ src/Andastra/Runtime/Graphics/Stride/Audio/StrideVoicePlayer.cs
@@ -697,7 +697,7 @@ namespace Andastra.Runtime.Stride.Audio
                     try
                     {
                         IntPtr bufferPtr = silenceHandle.AddrOfPinnedObject();
-                        FillBuffer(bufferPtr, bufferSizeBytes, (Stride.Audio.AudioLayer.BufferType)AudioLayer.BufferType.EndOfStream);
+                        FillBuffer(bufferPtr, bufferSizeBytes, (Stride.Audio.AudioLayer.BufferType)(int)AudioLayer.BufferType.EndOfStream);
                     }
                     finally
                     {
@@ -732,7 +732,7 @@ namespace Andastra.Runtime.Stride.Audio
             try
             {
                 IntPtr bufferPtr = bufferHandle.AddrOfPinnedObject();
-                FillBuffer(bufferPtr, bytesToCopy, (Stride.Audio.AudioLayer.BufferType)bufferType);
+                FillBuffer(bufferPtr, bytesToCopy, (Stride.Audio.AudioLayer.BufferType)(int)bufferType);
             }
             finally
             {
